move environment setup for wordpress and matomo in unit tests to new fixture classes
