+++ tests/phpunit/framework/test-case.php
@@ -7,33 +7,21 @@ use \WpMatomo\Capabilities;
 
 class MatomoUnit_TestCase extends WP_UnitTestCase {
 
+	/**
+	 * @var MatomoUnit_WordPress_Fixture
+	 */
+	private $wordpress_fixture;
+
 	public function setUp(): void {
 		parent::setUp();
 
-		if ( ! function_exists( 'wp_delete_site' ) ) {
-			function wp_delete_site( $site_id ) {
-				wpmu_delete_blog( $site_id, true );
-			}
-		}
+		$this->wordpress_fixture = new MatomoUnit_WordPress_Fixture();
+		$this->wordpress_fixture->set_up();
+	}
 
-		set_current_screen( 'front' );
-
-		$this->reset_roles();
-
-		add_filter(
-			'plugins_url',
-			function ( $url ) {
-				// workaround for https://github.com/wp-cli/wp-cli/issues/1037
-				// WP is installed in tmp dir, but our plugin must be symlinked in actual dir, then plugin_basename
-				// fails to remove the actual plugin path
-				// replaces eg http://example.org/wp-content/plugins/Users/foobar/www/wordpress-tests/src/wp-content/plugins/matomo/app/matomo.js
-				// with http://example.org/wp-content/plugins/matomo/app/matomo.js
-				return str_replace( rtrim( dirname( plugin_dir_path( MATOMO_ANALYTICS_FILE ) ), '/' ), '', $url );
-			}
-		);
-		if ( ! empty( $GLOBALS['wpdb'] ) ) {
-			$GLOBALS['wpdb']->suppress_errors( true );
-		}
+	public function tearDown(): void {
+		$this->wordpress_fixture->tear_down();
+		parent::tearDown();
 	}
 
 	protected function assume_admin_page() {
@@ -60,17 +48,6 @@ class MatomoUnit_TestCase extends WP_UnitTestCase {
 		return $id;
 	}
 
-	/**
-	 * Reset roles so they won't be stored across tests...
-	 */
-	protected function reset_roles() {
-		foreach ( array( 'editor', 'author', 'contributor' ) as $role ) {
-			get_role( $role )->remove_cap( Capabilities::KEY_SUPERUSER );
-			get_role( $role )->remove_cap( Capabilities::KEY_WRITE );
-			get_role( $role )->remove_cap( Capabilities::KEY_ADMIN );
-			get_role( $role )->remove_cap( Capabilities::KEY_VIEW );
-		}
-	}
 
 	/**
 	 * @return string
